Create bouncyBall.py
When the ball reaches the bottom of the window, its "y" changes to negative. When the ball reaches the top of the window, its "y" changes to positive.
When the ball reaches the right side of the window, its "x" changes to negative. When the ball reaches the left side of the window, its "x" changes to positive.